+++ src/Life/Github.hs
@@ -1,24 +1,29 @@
 -- | Utilities to work with GitHub repositories using "hub".
 
 module Life.Github
-       ( Owner (..)
-       , Repo  (..)
+       ( Owner  (..)
+       , Repo   (..)
 
          -- * Repository utils
        , checkRemoteSync
        , cloneRepo
        , insideRepo
+       , isBranchExists
        , withSynced
 
          -- * Repository manipulation commands
-       , CopyDirection (..)
-       , copyLife
        , addToRepo
+       , copyLife
+       , createNewBranch
        , createRepository
        , pullUpdateFromRepo
        , removeFromRepo
+       , setCurrentBranch
        , updateDotfilesRepo
        , updateFromRepo
+
+         -- * Constants
+       , master
        ) where
 
 import Control.Exception (catch, throwIO)
@@ -26,33 +31,34 @@ import Path (Abs, Dir, File, Path, Rel, toFilePath, (</>))
 import Path.IO (copyDirRecur, copyFile, getHomeDir, withCurrentDir)
 import System.IO.Error (IOError, isDoesNotExistError)
 
-import Life.Configuration (LifeConfiguration (..), lifeConfigMinus, parseRepoLife)
+import Life.Core (Owner(..), Repo(..), Branch (..), CopyDirection (..), CommitMsg (..))
+import Life.Configuration (LifeConfiguration (..), getBranch, lifeConfigMinus, parseRepoLife, parseHomeLife)
 import Life.Message (chooseYesNo, errorMessage, infoMessage, warningMessage)
 import Life.Shell (lifePath, relativeToHome, repoName, ($|))
 
-newtype Owner = Owner { getOwner :: Text } deriving (Show)
-newtype Repo  = Repo  { getRepo  :: Text } deriving (Show)
+import qualified Data.Text as T
 
 ----------------------------------------------------------------------------
 -- VSC commands
 ----------------------------------------------------------------------------
 
-askToPushka :: Text -> IO ()
-askToPushka commitMsg = do
+askToPushka :: Branch -> CommitMsg -> IO ()
+askToPushka branch@(Branch branchName) commitMsg = do
+    "git" ["checkout", branchName]
     "git" ["add", "."]
     infoMessage "The following changes are going to be pushed:"
     "git" ["diff", "--name-status", "HEAD"]
     continue <- chooseYesNo "Would you like to proceed?"
     if continue
-    then pushka commitMsg
+    then pushka commitMsg branch
     else errorMessage "Abort pushing" >> exitFailure
 
 -- | Make a commit and push it.
-pushka :: Text -> IO ()
-pushka commitMsg = do
+pushka :: CommitMsg -> Branch -> IO ()
+pushka (CommitMsg commitMsg) (Branch branch) = do
     "git" ["add", "."]
     "git" ["commit", "-m", commitMsg]
-    "git" ["push", "-u", "origin", "master"]
+    "git" ["push", "-u", "origin", branch]
 
 -- | Creates repository on GitHub inside given folder.
 createRepository :: Owner -> Repo -> IO ()
@@ -60,7 +66,7 @@ createRepository (Owner owner) (Repo repo) = do
     let description = ":computer: Configuration files"
     "git" ["init"]
     "hub" ["create", "-d", description, owner <> "/" <> repo]
-    pushka "Create the project"
+    pushka "Create the project" master
 
 ----------------------------------------------------------------------------
 -- dotfiles workflow
@@ -73,8 +79,8 @@ insideRepo action = do
     withCurrentDir repoPath action
 
 -- | Commits all changes inside 'repoName' and pushes to remote.
-pushRepo :: Text -> IO ()
-pushRepo = insideRepo . askToPushka
+pushRepo :: CommitMsg -> Branch -> IO ()
+pushRepo commitMsg branch = insideRepo $ askToPushka branch commitMsg
 
 -- | Clones @dotfiles@ repository assuming it doesn't exist.
 cloneRepo :: Owner -> IO ()
@@ -85,35 +91,54 @@ cloneRepo (Owner owner) = do
         "git" ["clone", "git@github.com:" <> owner <> "/dotfiles.git"]
 
 -- | Returns true if local @dotfiles@ repository is synchronized with remote repo.
-checkRemoteSync :: IO Bool
-checkRemoteSync = do
-    "git" ["fetch", "origin", "master"]
-    localHash  <- "git" $| ["rev-parse", "master"]
-    remoteHash <- "git" $| ["rev-parse", "origin/master"]
+checkRemoteSync :: Branch -> IO Bool
+checkRemoteSync (Branch branchName) = do
+    "git" ["fetch", "origin"]
+    localHash  <- "git" $| ["rev-parse", branchName]
+    remoteHash <- "git" $| ["rev-parse", "origin/" <> branchName]
     pure $ localHash == remoteHash
 
-withSynced :: IO a -> IO a
-withSynced action = insideRepo $ do
+withSynced :: Branch -> IO a -> IO a
+withSynced branch@(Branch branchName) action = insideRepo $ do
     infoMessage "Checking if repo is synchnorized..."
-    isSynced <- checkRemoteSync
-    if isSynced then do
+    isSyсnced <- checkRemoteSync branch
+    if isSyсnced then do
         infoMessage "Repo is up-to-date"
         action
-    else do
-        warningMessage "Local version of repository is out of date"
-        shouldSync <- chooseYesNo "Do you want to sync repo with remote?"
-        if shouldSync then do
-            "git" ["rebase", "origin/master"]
-            action
         else do
-            errorMessage "Aborting current command because repository is not synchronized with remote"
-            exitFailure
+            warningMessage "Local version of repository is out of date"
+            shouldSync <- chooseYesNo "Do you want to sync repo with remote?"
+            if shouldSync then do
+                setCurrentBranch branch
+                "git" ["rebase", "origin/" <> branchName]
+                action
+              else do
+                errorMessage "Aborting current command because repository is not synchronized with remote"
+                exitFailure
+
+-- | Try to set branch in @dotfiles@
+setCurrentBranch :: Branch  -> IO ()
+setCurrentBranch (Branch branchName) = insideRepo $ "git" ["checkout", branchName]
+
+-- | Check if branch exists
+isBranchExists :: Branch -> IO Bool
+isBranchExists (Branch branchName) = insideRepo $ do
+    "git" ["fetch", "origin"]
+    localBranches  <- "git" $| ["branch", "--list"]
+    remoteBranches <- "git" $| ["branch", "--list", "-r"]
+    pure $ elem branchName $ preformat localBranches ++ map cropRemotePrefixes (preformat remoteBranches)
+    where preformat = T.lines . T.filter (not . (\el -> el `elem` ['*', ' '])) . T.pack
+          cropRemotePrefixes raw = fromMaybe "" $ viaNonEmpty last $ T.splitOn "/" raw
+
+-- | Create new branch, switch to it and push it to Github
+createNewBranch :: Branch -> IO ()
+createNewBranch (Branch branchName) = insideRepo $ do
+    "git" ["checkout", "-b", branchName]
+    "git" ["push", "--set-upstream", "origin", branchName]
 
 ----------------------------------------------------------------------------
 -- File manipulation
 ----------------------------------------------------------------------------
-
-data CopyDirection = FromHomeToRepo | FromRepoToHome
 
 pullUpdateFromRepo :: LifeConfiguration -> IO ()
 pullUpdateFromRepo life = do
@@ -129,10 +154,10 @@ updateFromRepo excludeLife = insideRepo $ do
 
     copyLife FromRepoToHome lifeToLive
 
-updateDotfilesRepo :: Text -> LifeConfiguration -> IO ()
+updateDotfilesRepo :: CommitMsg -> LifeConfiguration -> IO ()
 updateDotfilesRepo commitMsg life = do
     copyLife FromHomeToRepo life
-    pushRepo commitMsg
+    pushRepo commitMsg (getBranch life)
 
 copyLife :: CopyDirection -> LifeConfiguration -> IO ()
 copyLife direction LifeConfiguration{..} = do
@@ -168,6 +193,8 @@ copyPathList copyAction direction pathList = do
 -- | Adds file or directory to the repository and commits
 addToRepo :: (Path Abs t -> Path Abs t -> IO ()) -> Path Rel t -> IO ()
 addToRepo copyFun path = do
+    life <- parseHomeLife
+
     -- copy file
     sourcePath <- relativeToHome path
     destinationPath <- relativeToHome (repoName </> path)
@@ -178,12 +205,13 @@ addToRepo copyFun path = do
     repoLifeFile <- relativeToHome (repoName </> lifePath)
     copyFile lifeFile repoLifeFile
 
-    let commitMsg = "Add: " <> toText (toFilePath path)
-    pushRepo commitMsg
+    let commitMsg = CommitMsg $ "Add: " <> toText (toFilePath path)
+    pushRepo commitMsg (getBranch life)
 
 -- | Removes file or directory from the repository and commits
 removeFromRepo :: (Path Abs t -> IO ()) -> Path Rel t -> IO ()
 removeFromRepo removeFun path = do
+    life <- parseHomeLife
     absPath <- relativeToHome (repoName </> path)
     catch (removeFun absPath) handleNotExist
 
@@ -192,8 +220,8 @@ removeFromRepo removeFun path = do
     repoLifeFile <- relativeToHome (repoName </> lifePath)
     copyFile lifeFile repoLifeFile
 
-    let commitMsg = "Remove: " <> pathTextName
-    pushRepo commitMsg
+    let commitMsg = CommitMsg $ "Remove: " <> pathTextName
+    pushRepo commitMsg (getBranch life)
   where
     pathTextName :: Text
     pathTextName = toText $ toFilePath path
@@ -202,3 +230,7 @@ removeFromRepo removeFun path = do
     handleNotExist e = if isDoesNotExistError e
         then errorMessage ("File/directory " <> pathTextName <> " is not found") >> exitFailure
         else throwIO e
+
+-- | Git "master" branch constant.
+master :: Branch
+master = Branch "master"