+++ src/Life/Main/Push.hs
@@ -10,9 +10,10 @@ import Lens.Micro.Platform ((^.))
 import Path (Abs, Path, Rel, toFilePath, (</>))
 import Path.IO (doesDirExist, doesFileExist, removeDirRecur, removeFile)
 
-import Life.Configuration (LifeConfiguration (..), directories, files, lifeConfigMinus,
-                           parseHomeLife, parseRepoLife)
-import Life.Github (updateDotfilesRepo, withSynced)
+import Life.Core (Branch (..))
+import Life.Configuration (LifeConfiguration (..), branch, getBranch,
+                           directories, files, lifeConfigMinus, parseHomeLife, parseRepoLife)
+import Life.Github (updateDotfilesRepo, withSynced, setCurrentBranch)
 import Life.Main.Init (lifeInitQuestion)
 import Life.Message (abortCmd)
 import Life.Shell (LifeExistence (..), relativeToHome, repoName, whatIsLife)
@@ -22,11 +23,13 @@ import qualified Data.Set as Set
 import qualified Data.Text as Text
 
 lifePush :: IO ()
-lifePush = whatIsLife >>= \case
-    OnlyRepo _ -> abortCmd "push" ".life file doesn't exist"
-    OnlyLife _ -> abortCmd "push" "dotfiles file doesn't exist"
-    NoLife     -> lifeInitQuestion "push" pushProcess
-    Both _ _   -> withSynced pushProcess
+lifePush = do
+    life <- parseHomeLife
+    whatIsLife >>= \case
+            OnlyRepo _ -> abortCmd "push" ".life file doesn't exist"
+            OnlyLife _ -> abortCmd "push" "dotfiles file doesn't exist"
+            NoLife     -> lifeInitQuestion "push" pushProcess
+            Both _ _   -> setCurrentBranch (getBranch life) >> withSynced (getBranch life) pushProcess
   where
     pushProcess :: IO ()
     pushProcess = do
@@ -55,6 +58,7 @@ lifePush = whatIsLife >>= \case
         pure $ LifeConfiguration
             <$> checkPaths eFiles
             <*> checkPaths eDirs
+            <*> (Success $ lf ^. branch :: Validation [Text] (Last Branch))
       where
         withExist :: (Path Abs f -> IO Bool) -> Path Rel f -> IO (Path Rel f, Bool)
         withExist doesExist path = (path,) <$> (relativeToHome path >>= doesExist)