+++ CHANGELOG.md
@@ -4,6 +4,19 @@ Change log
 life-sync uses [PVP Versioning][1].
 The change log is available [on GitHub][2].
 
+# 1.1.0
+
+* [#44](https://github.com/kowainik/life-sync/issues/44):
+  Use `Branch` for manage `~/dotfiles` branches.
+  Fix minor bug with broken `life add -d`.
+  Core module with internal types.
+  `life switch` branch management subcommand.
+
+# 1.0.2
+
+* [#42](https://github.com/kowainik/life-sync/issues/42):
+  Create newtype `Branch`.
+
 # 1.0.1
 
 * Switch to `relude`.